basic collision
this is what i was able to gather, essential the collision detection and the light curtain can be combined into one 'entity'. we do have the choice to do line plane or use multple spheres for the joints. with the line plane we can generate it, place a ply file of a human that crosses that and stops that or the robot itself. can also potentially used damepd least sqrs. will needto set some files as a robot so we can animate the human OR something else crossing the 'threshold'. lab 5-7 is where the stuff really is